+++ packages/testing/src/execution_testing/client_clis/clis/geth.py
@@ -54,6 +54,12 @@ class GethExceptionMapper(ExceptionMapper):
         TransactionException.PRIORITY_GREATER_THAN_MAX_FEE_PER_GAS: (
             "max priority fee per gas higher than max fee per gas"
         ),
+        TransactionException.TYPE_1_TX_PRE_FORK: (
+            "transaction type not supported"
+        ),
+        TransactionException.TYPE_2_TX_PRE_FORK: (
+            "transaction type not supported"
+        ),
         TransactionException.TYPE_3_TX_PRE_FORK: (
             "transaction type not supported"
         ),
@@ -95,6 +101,9 @@ class GethExceptionMapper(ExceptionMapper):
         BlockException.INVALID_GASLIMIT: "invalid gas limit",
         BlockException.INVALID_BLOCK_NUMBER: "invalid block number",
         BlockException.EXTRA_DATA_TOO_BIG: "invalid extradata length",
+        BlockException.INVALID_RECEIPTS_ROOT: "invalid receipt root hash",
+        BlockException.INVALID_LOG_BLOOM: "invalid bloom",
+        BlockException.INVALID_STATE_ROOT: "invalid merkle root",
     }
     mapping_regex: ClassVar[Dict[ExceptionBase, str]] = {
         TransactionException.TYPE_3_TX_MAX_BLOB_GAS_ALLOWANCE_EXCEEDED: (