+++ packages/testing/src/execution_testing/client_clis/clis/nimbus.py
@@ -114,5 +114,8 @@ class NimbusExceptionMapper(ExceptionMapper):
         BlockException.INVALID_GASLIMIT: "invalid gas limit",
         BlockException.INVALID_GAS_USED_ABOVE_LIMIT: "gasUsed should be non negative and smaller or equal gasLimit",
         BlockException.INVALID_BLOCK_HASH: "blockhash mismatch",
+        BlockException.INVALID_STATE_ROOT: "stateRoot mismatch",
+        BlockException.INVALID_RECEIPTS_ROOT: "receiptRoot mismatch",
+        BlockException.INVALID_LOG_BLOOM: "bloom mismatch",
     }
     mapping_regex: ClassVar[Dict[ExceptionBase, str]] = {}